% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{catch_error}
\alias{catch_error}
\title{Catch an error if reports are created}
\usage{
catch_error(snr, audience, ubb, results, ganztag, stype)
}
\arguments{
\item{snr}{Schoolnumber}

\item{audience}{Report audience}

\item{ubb}{UBB TRUE or FALSE}

\item{results}{Text string for Reports}

\item{ganztag}{Ganztag TRUE or FALSE}

\item{stype}{School type}
}
\value{
List with warnings and errors
}
\description{
The function runs a function (here: run_aslist) and captures
all errors and warnings.
}
