% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{catch_errors}
\alias{catch_errors}
\title{Catch all erros and warnings if several reports are created}
\usage{
catch_errors(mylist)
}
\arguments{
\item{mylist}{Schoolnumber}
}
\value{
List with warnings and errors
}
\description{
The function used the catch_error function to catch all errors
and warnings for a list of reports. Furthermore it writes the errors into
a log file.
}
