% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_plot}
\alias{check_plot}
\title{Check if variable/plot name exists}
\usage{
check_plot(user, server, credential, snr, year, ubb, audience, stype, ganztag)
}
\arguments{
\item{user}{User}

\item{server}{Server}

\item{credential}{Credential}

\item{snr}{Schoolnumber}

\item{year}{Year}

\item{ubb}{UBB TRUE or FALSE}

\item{audience}{Audience}

\item{stype}{School type}

\item{ganztag}{Ganztag}
}
\description{
The function retrieves plots from the meta data and compares them
with the names of plots listed in LimeSurvey.
}
