% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_plots}
\alias{check_plots}
\title{Run check_plot with a list of surveys}
\usage{
check_plots(list)
}
\arguments{
\item{list}{List (SNR, audience, ubb, etc.)}
}
\description{
The function is a wrapper to run check_plot for several schools
on the fly
}
