% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_response}
\alias{check_response}
\title{Check whether a survey has valid response}
\usage{
check_response(user, credential, server, snr, audience, ubb)
}
\arguments{
\item{user}{User}

\item{credential}{Credential}

\item{server}{Server}

\item{snr}{School number}

\item{audience}{Audience}

\item{ubb}{UBB TRUE or FALSE}
}
\description{
Check if a survey has valid response
}
