% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_templates}
\alias{check_templates}
\title{Check whether a report template exists}
\usage{
check_templates(user, credential, server)
}
\arguments{
\item{user}{User}

\item{credential}{Credential}

\item{server}{Server}
}
\description{
Check if a survey template can be found in meta data
}
