% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_folder}
\alias{create_folder}
\title{Create folder}
\usage{
create_folder(folderlist = c("res", "doc", "log", "orig", "prog"))
}
\arguments{
\item{folderlist}{Folderlist}
}
\description{
Create all folder for report
}
