% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{folder_exist}
\alias{folder_exist}
\title{Check if folder exists}
\usage{
folder_exist(folder)
}
\arguments{
\item{folder}{Foldername}
}
\value{
Foldername or warning
}
\description{
Helpfunction to check if a folder name exists (see create_folder)
}
