% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_list_surveys}
\alias{get_list_surveys}
\title{Get the list of all surveys from LimeSurvey}
\usage{
get_list_surveys(export = FALSE)
}
\arguments{
\item{export}{Export the list to an Excel file}
}
\description{
Get unique filter variable names
}
