% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{reports}
\alias{reports}
\title{The Report Data}
\format{
A data frame with 112 rows and 5 variables:
\describe{
\item{report}{The Report name}
\item{vars}{The variable name}
\item{plot}{The plot name}
\item{label}{The label}
\item{label_short}{The label short}
\item{sets}{The sets}
\item{type}{The type}
\item{timestamp}{Timestamp of creation}
...
}

A data frame with 112 rows and 5 variables:
\describe{
\item{stype}{The school type}
\item{ganztag}{Ganztag}
\item{type}{Survey type}
\item{surveys}{Survey template}
\item{rprtpckg}{Report package}
\item{ubb}{UBB}
\item{report_tmpl}{Report template}
\item{timestamp}{Timestamp of creation}
...
}
}
\description{
This data defines sets, lables, colors, etc.

This data defines sets, lables, colors, etc.
}
