% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run}
\alias{run}
\title{#' Superseded Create log
#' @description Create log file for a report.
#' @param snr Schoolnumber
#' @param year Year
#' @param audience Reporting group
#' @param logger tmp.logile}
\usage{
run(...)
}
\value{
Returns Rds, Plots, and PDF report
}
\description{
Run all functions to create a report at once
}
\details{
create_log <- function (snr,
year,
audience,
logger) {

if (logger == TRUE) {
tmp.dir_res <- paste0("res/", snr,"_", year, "/", audience, "/")

\if{html}{\out{<div class="sourceCode">}}\preformatted{#Del old logs if available
logfiles <- list.files(path = here::here(tmp.dir_res),
                       pattern = ".log", full.names = T)

if (length(logfiles) > 0) \{
  unlink(logfiles)
\}


#tmp.log.name <- paste0(lubridate::today(), "_", snr, "_")
log_file <- tempfile(pattern = "log_",
                     tmpdir = here::here(tmp.dir_res), fileext = ".log")
log <- log4r::logger(appenders = log4r::file_appender(log_file))
assign("log", value = log, envir=globalenv())
#log4r::info(log, glue::glue("Log for school: \{snr\}"))
}\if{html}{\out{</div>}}

}
}
Run
}
