% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_oes.R
\name{theme_oes}
\alias{theme_oes}
\title{Render OES Theme}
\usage{
theme_oes(base_size = 12, base_family = "Lato", device = "pdf")
}
\arguments{
\item{base_size}{An integer specifying a font size; defaults to \code{12}.}

\item{base_family}{A character string identifying a font family; defaults to
the OES-preferred \code{"Lato"}.}

\item{device}{A character string identifying a plotting device; defaults to
\code{"pdf"}.}
}
\value{
A list specifying a ggplot theme. See \code{Details} for more.
}
\description{
Defines the OES plotting theme
}
\details{
A list of length 93 of classes \code{theme} and \code{gg} that
defines elements of the display other than the data. \code{theme_oes()} wraps
\code{ggplot2::theme_bw()} and replaces several default values.
}
\author{
Miles Williams
}
